pyproject.toml
[tool.poetry]
name = "codepoint-api-rest"
version = "1.0.0"
description = "Uma api rest feita para atender ao desafio da codepoint, seguindo critérios e parâmetros estabelecidos pelo enunciado."
authors = ["angelo-francisco <ics20080729@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-ninja = "^1.3.0"
taskipy = "^1.14.1"
django-cors-headers = "^4.7.0"
pyjwt = "^2.10.1"

[tool.taskipy.tasks]
run = 'python manage.py runserver'
mkmg = 'python manage.py makemigrations'
mg = 'python manage.py migrate'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
